analyzer separation
