+++ ASTS/settings.py
@@ -41,7 +41,8 @@ INSTALLED_APPS = [
     'rest_framework.authtoken',
     #'rest_auth',
     'crawler',
-    'api'
+    'api',
+    'analyzer'
 ]
 
 MIDDLEWARE = [
